adjust list of items on WorkspaceHubInitialPage and InitialSettingsPage
