+++ src/styles/index.ts
@@ -3069,8 +3069,6 @@ const styles = (theme: ThemeColors) =>
         },
 
         workspaceSettingsSectionContainer: {
-            borderBottomWidth: 1,
-            borderBottomColor: theme.border,
             ...spacing.pt4,
         },
